+++ src/main/java/at/hannibal2/skyhanni/utils/ItemPriceUtils.kt
@@ -37,10 +37,10 @@ object ItemPriceUtils {
         pastRecipes: List<PrimitiveRecipe> = emptyList(),
     ): Double? {
         when (this) {
+            SKYBLOCK_COIN -> return 1.0
             NeuInternalName.GEMSTONE_COLLECTION -> return 0.0
             NeuInternalName.JASPER_CRYSTAL -> return 0.0
             NeuInternalName.RUBY_CRYSTAL -> return 0.0
-            NeuInternalName.SKYBLOCK_COIN -> return 1.0
             NeuInternalName.WISP_POTION -> return 20_000.0
             NeuInternalName.ENCHANTED_HAY_BLOCK -> return 7_776.0
             NeuInternalName.TIGHTLY_TIED_HAY_BALE -> return 1_119_744.0
