+++ src/main/java/at/hannibal2/skyhanni/utils/LorenzUtils.kt
@@ -99,7 +99,7 @@ object LorenzUtils {
     fun fillTable(
         data: List<DisplayTableEntry>,
         padding: Int = 1,
-        itemScale: Double = NeuItems.itemFontSize,
+        itemScale: Double = NeuItems.ITEM_FONT_SIZE,
     ): Renderable {
         val sorted = data.sortedByDescending { it.sort }
