+++ src/main/java/at/hannibal2/skyhanni/utils/NeuItems.kt
@@ -155,79 +155,7 @@ object NeuItems {
         return string.substring(prefix.length).toInternalName()
     }
 
-    const val itemFontSize = 2.0 / 3.0
-
-    fun ItemStack.renderOnScreen(
-        x: Float,
-        y: Float,
-        scaleMultiplier: Double = itemFontSize,
-        rescaleSkulls: Boolean = true,
-    ) {
-        val item = checkBlinkItem()
-        val isSkull = rescaleSkulls && item.item === Items.skull
-
-        val baseScale = (if (isSkull) 4f / 3f else 1f)
-        val finalScale = baseScale * scaleMultiplier
-
-        val translateX: Float
-        val translateY: Float
-        if (isSkull) {
-            val skullDiff = ((scaleMultiplier) * 2.5).toFloat()
-            translateX = x - skullDiff
-            translateY = y - skullDiff
-        } else {
-            translateX = x
-            translateY = y
-        }
-
-        GlStateManager.pushMatrix()
-
-        GlStateManager.translate(translateX, translateY, -19f)
-        GlStateManager.scale(finalScale, finalScale, 0.2)
-        GL11.glNormal3f(0f, 0f, 1f / 0.2f) // Compensate for z scaling
-
-        RenderHelper.enableGUIStandardItemLighting()
-
-        AdjustStandardItemLighting.adjust() // Compensate for z scaling
-
-        try {
-            Minecraft.getMinecraft().renderItem.renderItemIntoGUI(item, 0, 0)
-        } catch (e: Exception) {
-            if (lastWarn.passedSince() > 1.seconds) {
-                lastWarn = SimpleTimeMark.now()
-                println(" ")
-                println("item: $item")
-                println("name: ${item.displayName}")
-                println("getInternalNameOrNull: ${item.getInternalNameOrNull()}")
-                println(" ")
-                ChatUtils.debug("rendering an item has failed.")
-            }
-        }
-        RenderHelper.disableStandardItemLighting()
-
-        GlStateManager.popMatrix()
-    }
-
-    private var lastWarn = SimpleTimeMark.farPast()
-
-    private object AdjustStandardItemLighting {
-
-        private const val lightScaling = 2.47f // Adjust as needed
-        private const val g = 0.6f // Original Value taken from RenderHelper
-        private const val lightIntensity = lightScaling * g
-        private val itemLightBuffer = GLAllocation.createDirectFloatBuffer(16)
-
-        init {
-            itemLightBuffer.clear()
-            itemLightBuffer.put(lightIntensity).put(lightIntensity).put(lightIntensity).put(1.0f)
-            itemLightBuffer.flip()
-        }
-
-        fun adjust() {
-            GL11.glLight(16384, 4609, itemLightBuffer)
-            GL11.glLight(16385, 4609, itemLightBuffer)
-        }
-    }
+    const val ITEM_FONT_SIZE = 2.0 / 3.0
 
     fun allNeuRepoItems(): Map<String, JsonObject> = EnoughUpdatesManager.getItemInformation()
