+++ versions/1.21.4/src/main/java/at/hannibal2/skyhanni/utils/NeuCalculator.kt
@@ -1,0 +1,10 @@
+package at.hannibal2.skyhanni.utils
+
+import java.math.BigDecimal
+
+object NeuCalculator {
+
+    fun calculateOrNull(input: String): BigDecimal? {
+        return null
+    }
+}
